app/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests==2.22.0
pydantic==1.8.2
python-dotenv==0.19.0